+++ pkg/analysis_server/lib/src/services/correction/dart/add_missing_required_argument.dart
@@ -99,8 +99,7 @@ class AddMissingRequiredArgument extends ResolvedCorrectionProducer {
       var codeStyleOptions = getCodeStyleOptions(unitResult.file);
       var defaultValue = getDefaultStringParameterValue(
           missingParameter, codeStyleOptions,
-          withNullability: libraryElement.isNonNullableByDefault &&
-              (missingParameter.library?.isNonNullableByDefault ?? false));
+          withNullability: true);
 
       await builder.addDartFileEdit(file, (builder) {
         builder.addInsertion(offset, (builder) {