+++ pkg/analyzer/lib/src/dart/analysis/library_analyzer.dart
@@ -488,7 +488,6 @@ class LibraryAnalyzer {
     if (sdkVersionConstraint != null) {
       SdkConstraintVerifier verifier = SdkConstraintVerifier(
         errorReporter,
-        _libraryElement,
         sdkVersionConstraint.withoutPreRelease,
       );
       unit.accept(verifier);
@@ -809,7 +808,7 @@ class LibraryAnalyzer {
       for (var i = 0; i < docImports.length; i++) {
         _resolveLibraryDocImportDirective(
           directive: docImports[i].import as ImportDirectiveImpl,
-          state: containerKind.libraryDocImports[i],
+          state: containerKind.docImports[i],
           errorReporter: containerErrorReporter,
         );
       }