+++ pkg/analyzer/test/id_tests/inheritance_test.dart
@@ -28,16 +28,27 @@ main(List<String> args) async {
         runTestFor(const _InheritanceDataComputer(), [analyzerDefaultConfig]),
     skipList: [
       // Legacy, not supported by the analyzer anymore.
+      'covariant_opt_out.dart',
       'from_opt_in',
       'from_opt_out',
+      'generic_members_from_opt_in',
+      'generic_members_from_opt_out',
       'in_out_in',
       'infer_from_opt_in',
+      'infer_opt_in_from_mixed',
       'infer_opt_out_from_mixed',
       'infer_parameter_opt_out.dart',
       'issue40414',
+      'issue40481',
+      'issue40524',
+      'issue40553',
       'member_from_opt_in',
+      'member_from_opt_out',
       'members_from_opt_in',
       'members_from_opt_out',
+      'members_opt_out.dart',
+      'nsm_from_opt_in',
+      'sink.dart',
       'top_merge_opt_out.dart',
     ],
     skipMap: {
@@ -97,7 +108,7 @@ class _InheritanceDataExtractor extends AstDataExtractor<String> {
   @override
   String? computeElementValue(Id id, Element element) {
     if (element is LibraryElement) {
-      return 'nnbd=${element.isNonNullableByDefault}';
+      return 'nnbd=true';
     }
     return null;
   }